backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
jinja2>=3.1
pydantic>=2.6
httpx>=0.27
anthropic>=0.40
python-dotenv>=1.0
google-cloud-firestore>=2.15
